requirements.txt
asgiref==3.2.3
boto3==1.10.34
botocore==1.13.34
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.8
django-cors-headers==3.2.0
django-filter==2.2.0
django-rest-auth==0.9.5
django-tenant-schemas==1.10.0
djangorestframework==3.10.3
docutils==0.15.2
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
ordered-set==3.1.1
packaging==19.2
psycopg2-binary==2.8.4
pyparsing==2.4.5
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
